Added configurable setting for international dialing
The default configuration for the polycom phones is having international
dialing enabled which causes the star button to also support dialing +
for international calls.  This means when using certain feature codes
that start with ** the user must wait a few seconds in-between presses
which can be problematic.  Now they can simply disable with the
polycom_international_enabled setting set to 0 to disable.